requirements.txt
langchain
faiss-cpu
streamlit
llama-index
ollama